Implement MVC paradigm (#2)
- Models are event-emitting classes, which emit "update" events when
updated via 'setter' methods.
- Views are subclasses of widgets. A model is passed in to the
constructor, and the constructor will instantiate a controller. Views
are just responsible for displaying widgets (or other content), and do
not themselves react to widget (or other UI) events, nor model update
events.
- Controllers are plain classes, which are passed a view widget and a
model in their constructor. The controller updates the view from the model
data whenever there is an "update" event from the model. The controller
also listens to relevant events from the widgets, and calls the
appropriate setter method on the model.

To make the code testable, the globals `$`, `mw`, and `OO` are defined
in new globals.js file, which is automatically built to either pass
through values from the window object (when building code to run in the
browser), or imported from packages -- either Node module imports, or
a mock file in the \mocks directory if there is no existing module with
a suitable interface.

Mocha is used for testing, with tests in the \test directory.