[WIP] Translation / i18n support
Added the basic mechanism for having messages stored in a json file
on-wiki, loaded into a private instance of jQuery.i18n (set to the
locale of the wiki, not the user), and usable within the source files.

Banana-i18n is used to provide a mock jQuery.i18n instance for
development/testing purposes.