build: update browser targets
### What

Looks like the transpiling works by feeding it browser version numbers that you want to support. This is done in the .babelrc file. This file hasn't been touched in 5 years and the version numbers can be bumped way up. Bump them up to match the MediaWiki Grade A browsers table. As part of this, support for Internet Explorer is dropped in favor of Edge.

### Why

- Should reduce the number of polyfills needed, reducing the build size.